Refcatored timer to logger package
