--- packages/core/test/unit/timing.test.js
@@ -1,40 +1,0 @@
-import TimeIt from '../../src/timing.js';
-import { logger as mockLogger, setupTest } from '@percy/cli-command/test/helpers';
-import logger from '@percy/logger';
-
-describe('TimeIt', () => {
-  beforeEach(async () => {
-    await setupTest();
-  });
-
-  afterEach(() => {
-  });
-
-  it('should return same instance everytime', () => {
-    const obj1 = new TimeIt();
-    const obj2 = new TimeIt();
-    expect(obj1).toEqual(obj2);
-  });
-
-  describe('measure', () => {
-    it('should execute callback and log duration', async () => {
-      const log = logger('test');
-      const date1 = new Date(2024, 4, 11, 13, 30, 0);
-      const date2 = new Date(2024, 4, 11, 13, 31, 0);
-      const meta = { abc: '123' };
-      // Logger internall calls Date.now, so need to mock
-      // response for it as well.
-      spyOn(Date, 'now').and.returnValues(date1, date1, date2, date1);
-      const timeit = new TimeIt();
-      timeit.log.loglevel('debug');
-      const callback = async () => { log.info('abcd'); };
-      await timeit.measure('step', 'test', meta, callback);
-      expect(mockLogger.stdout).toEqual([
-        '[percy] abcd'
-      ]);
-      expect(mockLogger.stderr).toEqual([
-        '[percy:timer] step - test - 60s'
-      ]);
-    });
-  });
-});
